MVP-6 Naht 1: RPE-Welle neu (ankert rpe_low→rpe_high, 0.5-Raster)
Welle war nach Rundung platt (alle Wochen rundeten auf denselben int).
Neu: float-RPE ankert an rpe_low, rampt auf rpe_high; Progression = Mitte
(0.5er bei L1/L2, 1.0er bei L3/L4). Deload = rpe_low−1 (Floor 4), damit
nicht RPE-gleich zur Akku-Woche. Volumen bleibt flach (intensitätsgeführt).
RPE int→float (Frontend-Vertrag, :g-Format nur im PDF). 5 neue Wellen-Asserts.

+++ claude/claude_client.py
@@ -32,7 +32,7 @@ def generiere_uebungsauswahl(
     uebungen_gefiltert: dict[str, list[dict]],
     woche_typ: str,
     ziel_saetze: int,
-    ziel_rpe: int,
+    ziel_rpe: float,
 ) -> ClaudeOutput:
     client = anthropic.Anthropic(api_key=os.environ["ANTHROPIC_API_KEY"])
